Split the generation and creation of named object instances out of `DefaultObjectFactory`. This separation will allow `DefaultObjectFactory` to later be contextualised with scoped services to be injected into the objects that it creates, as currently it is a global service due to the generation of named object types.
