+++ subprojects/model-core/src/test/groovy/org/gradle/api/internal/model/DefaultObjectFactoryTest.groovy
@@ -16,14 +16,8 @@
 
 package org.gradle.api.internal.model
 
-import org.gradle.api.GradleException
 import org.gradle.api.Named
 import org.gradle.test.fixtures.concurrent.ConcurrentSpec
-import org.gradle.util.Matchers
-import spock.lang.Ignore
-
-import java.util.concurrent.CopyOnWriteArrayList
-import java.util.concurrent.CyclicBarrier
 
 class DefaultObjectFactoryTest extends ConcurrentSpec {
     def factory = DefaultObjectFactory.INSTANCE
@@ -38,203 +32,6 @@ class DefaultObjectFactoryTest extends ConcurrentSpec {
 
         n1.name == "a"
         n2.name == "b"
-
-        Matchers.strictlyEquals(n1, n1)
-        n2 != n1
-        !n2.equals(n1)
-
-        n1.hashCode() == n1.hashCode()
-        n1.hashCode() != n2.hashCode()
-
-        n1.toString() == "a"
-        n2.toString() == "b"
-
-        n1.is(factory.named(Named, "a"))
-        n2.is(factory.named(Named, "b"))
-    }
-
-    def "creates instance of subtype of Named"() {
-        expect:
-        def n1 = factory.named(CustomNamed, "a")
-        def n2 = factory.named(CustomNamed, "b")
-
-        n1.is(n1)
-        !n1.is(n2)
-
-        n1.name == "a"
-        n2.name == "b"
-
-        Matchers.strictlyEquals(n1, n1)
-        n2 != n1
-        !n2.equals(n1)
-
-        n1.hashCode() == n1.hashCode()
-        n1.hashCode() != n2.hashCode()
-
-        n1.toString() == "a"
-        n2.toString() == "b"
-
-        n1.is(factory.named(CustomNamed, "a"))
-        n2.is(factory.named(CustomNamed, "b"))
-
-        !n1.is(factory.named(Named, "a"))
-        !n2.is(factory.named(Named, "b"))
-    }
-
-    def "multiple threads can create different instances of same type"() {
-        def results = new CopyOnWriteArrayList()
-        def barrier = new CyclicBarrier(10)
-
-        when:
-        async {
-            10.times { n ->
-                start {
-                    barrier.await()
-                    results.add(factory.named(CustomNamed, n as String))
-                }
-            }
-        }
-
-        then:
-        results.size() == 10
-        results.unique().size() == 10
-    }
-
-    def "multiple threads can create same instance"() {
-        def results = new CopyOnWriteArrayList()
-        def barrier = new CyclicBarrier(10)
-
-        when:
-        async {
-            10.times { n ->
-                start {
-                    barrier.await()
-                    results.add(factory.named(CustomNamed, "value"))
-                }
-            }
-        }
-
-        then:
-        results.size() == 10
-        results.unique().size() == 1
-    }
-
-    def "creates instance of abstract Java class"() {
-        expect:
-        def n1 = factory.named(AbstractNamed, "a")
-        def n2 = factory.named(AbstractNamed, "b")
-
-        n1.is(n1)
-        !n1.is(n2)
-
-        n1.name == "a"
-        n2.name == "b"
-
-        n1.calculatedValue == "[a]"
-        n2.calculatedValue == "[b]"
-
-        Matchers.strictlyEquals(n1, n1)
-        n2 != n1
-        !n2.equals(n1)
-
-        n1.hashCode() == n1.hashCode()
-        n1.hashCode() != n2.hashCode()
-
-        n1.toString() == "a"
-        n2.toString() == "b"
-
-        n1.is(factory.named(AbstractNamed, "a"))
-        n2.is(factory.named(AbstractNamed, "b"))
-
-        !n1.is(factory.named(Named, "a"))
-        !n2.is(factory.named(Named, "b"))
-
-        AbstractNamed.counter == 2
-    }
-
-    def "creates instance of Java class with dummy getName() implementation"() {
-        expect:
-        def n1 = factory.named(DummyNamed, "a")
-        def n2 = factory.named(DummyNamed, "b")
-
-        n1.is(n1)
-        !n1.is(n2)
-
-        n1.name == "a"
-        n2.name == "b"
-
-        n1.calculatedValue == "[a]"
-        n2.calculatedValue == "[b]"
-
-        Matchers.strictlyEquals(n1, n1)
-        n2 != n1
-        !n2.equals(n1)
-
-        n1.hashCode() == n1.hashCode()
-        n1.hashCode() != n2.hashCode()
-
-        n1.toString() == "a"
-        n2.toString() == "b"
-
-        n1.is(factory.named(DummyNamed, "a"))
-        n2.is(factory.named(DummyNamed, "b"))
-
-        !n1.is(factory.named(Named, "a"))
-        !n2.is(factory.named(Named, "b"))
-
-        DummyNamed.counter == 2
-    }
-
-    def "creates instance of Groovy class with dummy getName() implementation"() {
-        expect:
-        def n1 = factory.named(DummyGroovyNamed, "a")
-        def n2 = factory.named(DummyGroovyNamed, "b")
-
-        n1.is(n1)
-        !n1.is(n2)
-
-        n1.name == "a"
-        n2.name == "b"
-
-        n1.calculatedValue == "[a]"
-        n2.calculatedValue == "[b]"
-
-        Matchers.strictlyEquals(n1, n1)
-        n2 != n1
-        !n2.equals(n1)
-
-        n1.hashCode() == n1.hashCode()
-        n1.hashCode() != n2.hashCode()
-
-        n1.toString() == "a"
-        n2.toString() == "b"
-
-        n1.is(factory.named(DummyGroovyNamed, "a"))
-        n2.is(factory.named(DummyGroovyNamed, "b"))
-
-        !n1.is(factory.named(Named, "a"))
-        !n2.is(factory.named(Named, "b"))
-    }
-
-    def "class should not have any instance fields"() {
-        when:
-        factory.named(NamedWithFields, "a")
-
-        then:
-        def e = thrown(GradleException)
-        e.message == """Type ${NamedWithFields.name} is not a valid Named implementation class:
-- Field name is not valid: A Named implementation class must not define any instance fields.
-- Field other is not valid: A Named implementation class must not define any instance fields."""
-
-        when:
-        factory.named(NamedWithFields.Sub, "a")
-
-        then:
-        e = thrown(GradleException)
-        e.message == """Type ${NamedWithFields.name}.Sub is not a valid Named implementation class:
-- Field NamedWithFields.name is not valid: A Named implementation class must not define any instance fields.
-- Field NamedWithFields.other is not valid: A Named implementation class must not define any instance fields.
-- Field b is not valid: A Named implementation class must not define any instance fields."""
     }
 
     def "creates instance with constructor parameters"() {
@@ -244,16 +41,6 @@ class DefaultObjectFactoryTest extends ConcurrentSpec {
         def result = factory.newInstance(SomeType, param)
         result instanceof SomeType
         result.result == param
-    }
-
-    @Ignore
-    def "interface may not have additional methods"() {
-        expect: false
-    }
-
-    @Ignore
-    def "abstract class may not have additional abstract methods"() {
-        expect: false
     }
 }
